Add grace period before deleting expired pause, always skip

+++ pkg/execution/executor/executor.go
@@ -58,7 +58,8 @@ var (
 	// step should be safely retried.
 	ErrHandledStepError = fmt.Errorf("handled step error")
 
-	PauseHandleConcurrency = 100
+	PauseHandleConcurrency          = 100
+	PauseExpiredDeletionGracePeriod = time.Second * 10
 )
 
 var (
@@ -1451,9 +1452,15 @@ func (e *executor) handlePausesAllNaively(ctx context.Context, iter state.PauseI
 			// time to process.  We handle that here and assume that the event
 			// did not occur in time.
 			if pause.Expires.Time().Before(time.Now()) {
-				// Consume this pause to remove it entirely
-				l.Debug().Msg("deleting expired pause")
-				_ = e.pm.DeletePause(context.Background(), *pause)
+				l.Debug().Msg("encountered expired pause")
+
+				shouldDelete := pause.Expires.Time().Add(PauseExpiredDeletionGracePeriod).Before(time.Now())
+				if shouldDelete {
+					// Consume this pause to remove it entirely
+					l.Debug().Msg("deleting expired pause")
+					_ = e.pm.DeletePause(context.Background(), *pause)
+				}
+
 				return
 			}
 
@@ -1626,10 +1633,16 @@ func (e *executor) handleAggregatePauses(ctx context.Context, evt event.TrackedE
 			// time to process.  We handle that here and assume that the event
 			// did not occur in time.
 			if pause.Expires.Time().Before(time.Now()) {
-				// Consume this pause to remove it entirely
-				l.Debug("deleting expired pause")
-				_ = e.pm.DeletePause(context.Background(), pause)
-				_ = e.exprAggregator.RemovePause(ctx, pause)
+				l.Debug("encountered expired pause")
+
+				shouldDelete := pause.Expires.Time().Add(PauseExpiredDeletionGracePeriod).Before(time.Now())
+				if shouldDelete {
+					// Consume this pause to remove it entirely
+					l.Debug("deleting expired pause")
+					_ = e.pm.DeletePause(context.Background(), pause)
+					_ = e.exprAggregator.RemovePause(ctx, pause)
+				}
+
 				return
 			}
 
@@ -1743,9 +1756,15 @@ func (e *executor) HandleInvokeFinish(ctx context.Context, evt event.TrackedEven
 	}
 
 	if pause.Expires.Time().Before(time.Now()) {
-		// Consume this pause to remove it entirely
-		l.Debug().Msg("deleting expired pause")
-		_ = e.pm.DeletePause(context.Background(), *pause)
+		l.Debug().Msg("encountered expired pause")
+
+		shouldDelete := pause.Expires.Time().Add(PauseExpiredDeletionGracePeriod).Before(time.Now())
+		if shouldDelete {
+			// Consume this pause to remove it entirely
+			l.Debug().Msg("deleting expired pause")
+			_ = e.pm.DeletePause(context.Background(), *pause)
+		}
+
 		return nil
 	}
